requirements.txt
fastapi
uvicorn
python-multipart
pillow
torch
torchvision
matplotlib
sqlalchemy
psycopg2-binary 